Move result to libtock-rs
